feat: extract sign-in infrastructure into composable setup task, register OAuth strategies
- New `ash_authentication_phoenix.setup` task extracts sign-in infrastructure
  (AuthController, routes, pipeline plugs, LiveUserAuth, overrides, tailwind)
  from the install task into a standalone, composable task
- Install task now composes with setup instead of inlining the logic,
  eliminating code duplication for security-critical auth infrastructure
- Register all 8 OAuth/OIDC strategies in the AAP dispatcher, pointing to
  the setup task for Phoenix-side integration
- OAuth strategies auto-render as buttons in the sign-in page via the
  strategy's `icon` field — no provider-specific Phoenix components needed
